requirements.txt
streamlit==1.31.0
langchain==0.3.14
langchain-google-genai==2.1.4
langgraph==0.2.64
torpy==1.1.6
requests==2.31.0
beautifulsoup4==4.12.3
chromadb==0.4.22
python-dotenv==1.0.0
rich==13.7.0
networkx==3.2.1
matplotlib==3.8.2